changed float to double
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@3938 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/srccontrols/color/iup_colorhsi.c
@@ -10,17 +10,17 @@
 #include <stdio.h>
 #include "iup_colorhsi.h"
 
-static const float rad60 =  1.0471975f;
-static const float rad120 = 2.0943951f;
-static const float rad180 = 3.1415926f;
-static const float rad240 = 4.1887902f;
-static const float rad300 = 5.2359877f;
-static const float rad360 = 6.2831853f;
-static const float sqrt3 = 1.7320508f;
-static const float rad2deg = 57.2957795131f;
+static const double rad60 =  1.0471975;
+static const double rad120 = 2.0943951;
+static const double rad180 = 3.1415926;
+static const double rad240 = 4.1887902;
+static const double rad300 = 5.2359877;
+static const double rad360 = 6.2831853;
+static const double sqrt3 = 1.7320508;
+static const double rad2deg = 57.2957795131;
 
-static float costab[361];
-static float sintab[361];
+static double costab[361];
+static double sintab[361];
 static int init_tab = 0;
 
 static void iColorBuildTables(void)
@@ -28,19 +28,19 @@ static void iColorBuildTables(void)
   int theta;
   for (theta=0; theta<=360; theta++)
   {
-    float th = ((float)theta)/rad2deg;
-    costab[theta] = cosf(th);
-    sintab[theta] = sinf(th);
+    double th = ((double)theta)/rad2deg;
+    costab[theta] = cos(th);
+    sintab[theta] = sin(th);
   }
   init_tab = 1;
 }
 
-static void iColorSinCos(float H, float *sinH, float *cosH)
+static void iColorSinCos(double H, double *sinH, double *cosH)
 {
   int theta;
 
   H *= rad2deg;
-  theta = (int)(H + 0.5f);  /* Round */
+  theta = (int)(H + 0.5);  /* Round */
 
   if (!init_tab)
     iColorBuildTables();
@@ -52,32 +52,32 @@ static void iColorSinCos(float H, float *sinH, float *cosH)
   *sinH = sintab[theta];
 }
 
-static float iColorNormHue(float H)
+static double iColorNormHue(double H)
 {
-  while (H < 0.0f) 
+  while (H < 0.0) 
     H += rad360;
 
   if (H > rad360)  
-    H = fmodf(H, rad360);
+    H = fmod(H, rad360);
 
   return H;
 }
 
-static unsigned char iColorQuantize(float value)
+static unsigned char iColorQuantize(double value)
 {
-  if (value >= 1.0f) return 255;
-  if (value <= 0.0f) return 0;
+  if (value >= 1.0) return 255;
+  if (value <= 0.0) return 0;
   return (unsigned char)(value*256);  
 }                               
 
-static float iColorReconstruct(unsigned char value)
+static double iColorReconstruct(unsigned char value)
 {
-  if (value <= 0) return 0.0f;
-  if (value >= 255) return 1.0f;
-  return (((float)value + 0.5f)/256.0f);
+  if (value <= 0) return 0.0;
+  if (value >= 255) return 1.0;
+  return (((double)value + 0.5)/256.0);
 }
 
-static void iColorSmax01(float h, float hr, float hb, float hg, float *h0, float *h1)
+static void iColorSmax01(double h, double hr, double hb, double hg, double *h0, double *h1)
 {
   if (h < rad60)
   {
@@ -112,104 +112,104 @@ static void iColorSmax01(float h, float hr, float hb, float hg, float *h0, float
 }
 
 /* Given H and I, returns S max, but s is in u,v space. */
-static float iColorHSI_Smax(float h, float cosH, float sinH, float i)
+static double iColorHSI_Smax(double h, double cosH, double sinH, double i)
 {
-  float hr, hb, hg, imax, h0, h1;
+  double hr, hb, hg, imax, h0, h1;
 
   /* i here is normalized between 0-1 */
 
-  if (i == 0.0f || i == 1.0f)
-    return 0.0f;
+  if (i == 0.0 || i == 1.0)
+    return 0.0;
 
   /* Making r=0, g=0, b=0, r=1, g=1 or b=1 in the parametric equations and 
      writting s in function of H and I. */
 
   /* at bottom */
-  if (i <= 1.0f/3.0f)
+  if (i <= 1.0/3.0)
   {
     /* face B=0 */
     if (h < rad120)
     {
-      hb = (cosH + sinH*sqrt3)/3.0f;
+      hb = (cosH + sinH*sqrt3)/3.0;
       return i/hb;
     }
 
     /* face R=0 */
     if (h < rad240)
     {
-      hr = -cosH/1.5f;
+      hr = -cosH/1.5;
       return i/hr;
     }
 
     /* face G=0 (h < rad360) */
     {
-      hg = (cosH - sinH*sqrt3)/3.0f;
+      hg = (cosH - sinH*sqrt3)/3.0;
       return i/hg;
     }
   }
 
   /* at top */
-  if (i >= 2.0f/3.0f)
+  if (i >= 2.0/3.0)
   {
     /* face R=1 */
     if (h < rad60 || h > rad300)
     {
-      hr = cosH/1.5f;
-      return (1.0f-i)/hr;
+      hr = cosH/1.5;
+      return (1.0-i)/hr;
     }
 
     /* face G=1 */
     if (h < rad180)
     {
-      hg = (-cosH + sinH*sqrt3)/3.0f;
-      return (1.0f-i)/hg;
+      hg = (-cosH + sinH*sqrt3)/3.0;
+      return (1.0-i)/hg;
     }
 
     /* face B=1 (h > rad180 && h < rad300) */
     {
-      hb = (-cosH - sinH*sqrt3)/3.0f;
-      return (1.0f-i)/hb;
+      hb = (-cosH - sinH*sqrt3)/3.0;
+      return (1.0-i)/hb;
     }
   }
 
   /* in the middle */
-  hr = cosH/1.5f;
-  hg = (-cosH + sinH*sqrt3)/3.0f;
-  hb = (-cosH - sinH*sqrt3)/3.0f;
+  hr = cosH/1.5;
+  hg = (-cosH + sinH*sqrt3)/3.0;
+  hb = (-cosH - sinH*sqrt3)/3.0;
 
   iColorSmax01(h, hr, hb, hg, &h0, &h1);
 
-  if (h == 0.0f || h == rad120 || h == rad240)
-    imax = 1.0f/3.0f;
+  if (h == 0.0 || h == rad120 || h == rad240)
+    imax = 1.0/3.0;
   else if (h == rad60 || h == rad180 || h == rad300)
-    imax = 2.0f/3.0f;
+    imax = 2.0/3.0;
   else
     imax = h0/(h0 - h1);
 
   if (i < imax) 
     return -i/h0;
   else
-    return (1.0f-i)/h1;
+    return (1.0 - i)/h1;
 }
 
 /* Given H, returns I where S is max,
    BUT the maximum S here is 1 at the corners of the cube. */
-static float iColorHSI_ImaxS(float h, float cosH, float sinH)
+static double iColorHSI_ImaxS(double h, double cosH, double sinH)
 {
-  float i, h0, h1;
-  float hr, hb, hg;
+  double i, h0, h1;
+  double hr, hb, hg;
 
   /* i here is normalized between 0-1 */
 
-  if (h == 0.0f || h == rad120 || h == rad240)
-    return 1.0f/3.0f;
+  if (h == 0.0 || h == rad120 || h == rad240)
+    return 1.0/3.0;
 
   if (h == rad60 || h == rad180 || h == rad300)
-    return 2.0f/3.0f;
+    return 2.0/3.0;
 
-  hr = cosH/1.5f;
-  hg = (-cosH + sinH*sqrt3)/3.0f;
-  hb = (-cosH - sinH*sqrt3)/3.0f;
+  hr = cosH/1.5;
+  hg = (-cosH + sinH*sqrt3)/3.0;
+  hb = (-cosH - sinH*sqrt3)/3.0;
 
   iColorSmax01(h, hr, hb, hg, &h0, &h1);
 
@@ -218,29 +218,29 @@ static float iColorHSI_ImaxS(float h, float cosH, float sinH)
   return i;
 }
 
-static void iColorRGB2HSI(float r, float g, float b, float *h, float *s, float *i)
+static void iColorRGB2HSI(double r, double g, double b, double *h, double *s, double *i)
 {            
-  float v, u, ImaxS; 
+  double v, u, ImaxS; 
 
   /* Parametric equations */
-  v = r - (g + b)/2.0f;
-  u = (g - b) * (sqrt3/2.0f);
+  v = r - (g + b)/2.0;
+  u = (g - b) * (sqrt3/2.0);
 
   *i = (r + g + b)/3;  /* already normalized to 0-1 */
 
-  *s = sqrtf(v*v + u*u);  /* s is between 0-1, BUT it is linear in the cube and it is in u,v space. */
+  *s = sqrt(v*v + u*u);  /* s is between 0-1, BUT it is linear in the cube and it is in u,v space. */
 
-  if (*s == 0.0f)
+  if (*s == 0.0)
   {
     /*  *h = <any>  (left unchanged) */
-    ImaxS = 1.0f/3.0f;
+    ImaxS = 1.0/3.0;
   }
   else
   {
-    float Smax; 
-    float H, cosH, sinH;
+    double Smax; 
+    double H, cosH, sinH;
 
-    H = atan2f(u, v);
+    H = atan2(u, v);
     H = iColorNormHue(H);
     *h = H * rad2deg;
 
@@ -248,8 +248,8 @@ static void iColorRGB2HSI(float r, float g, float b, float *h, float *s, float *
 
     /* must scale S from 0-Smax to 0-1 */
     Smax = iColorHSI_Smax(H, cosH, sinH, *i);
-    if (Smax == 0.0f)
-      *s = 0.0f;
+    if (Smax == 0.0)
+      *s = 0.0;
     else
     {
       if (*s > Smax) /* because of round problems when calculating s and Smax */
@@ -257,20 +257,20 @@ static void iColorRGB2HSI(float r, float g, float b, float *h, float *s, float *
       *s /= Smax;
     }
 
-    ImaxS = iColorHSI_ImaxS((float)H, cosH, sinH);
+    ImaxS = iColorHSI_ImaxS((double)H, cosH, sinH);
   }
 
   /* must convert I from linear scale to non-linear scale. USED ONLY FOR THE COLORBROWSER */
-  if (*i < 0.5f) /* half I is I=ImaxS, not I=0.5 */
-    *i = ((*i)*0.5f)/ImaxS;
+  if (*i < 0.5) /* half I is I=ImaxS, not I=0.5 */
+    *i = ((*i)*0.5)/ImaxS;
   else
-    *i = (((*i) - ImaxS)*0.5f)/(1.0f - ImaxS) + 0.5f;
+    *i = (((*i) - ImaxS)*0.5)/(1.0 - ImaxS) + 0.5;
 }
 
-static void iColorHSI2RGB(float h, float s, float i, float *r, float *g, float *b)
+static void iColorHSI2RGB(double h, double s, double i, double *r, double *g, double *b)
 {
-  float cosH, sinH, H, v, u;
-  float Smax, ImaxS; 
+  double cosH, sinH, H, v, u;
+  double Smax, ImaxS; 
 
   if (i < 0) i = 0;
   else if (i > 1) i = 1;
@@ -278,7 +278,7 @@ static void iColorHSI2RGB(float h, float s, float i, float *r, float *g, float *
   if (s < 0) s = 0;
   else if (s > 1) s = 1;
 
-  if (s == 0.0f || i == 1.0f || i == 0.0f || (int)h == 360)
+  if (s == 0.0 || i == 1.0 || i == 0.0 || h == 360.0)
   {
     *r = i;
     *g = i;
@@ -293,49 +293,49 @@ static void iColorHSI2RGB(float h, float s, float i, float *r, float *g, float *
     
   /* must convert I from non-linear scale to linear scale. USED ONLY FOR THE COLORBROWSER */
   ImaxS = iColorHSI_ImaxS(H, cosH, sinH);
-  if (i < 0.5f) /* half I is I=ImaxS, not I=0.5 */
-    i  = ImaxS * (i / 0.5f);
+  if (i < 0.5) /* half I is I=ImaxS, not I=0.5 */
+    i  = ImaxS * (i / 0.5);
   else
-    i = (1.0f - ImaxS) * (i - 0.5f)/0.5f + ImaxS;
+    i = (1.0 - ImaxS) * (i - 0.5)/0.5 + ImaxS;
 
   /* must scale S from 0-1 to 0-Smax */
   Smax = iColorHSI_Smax(H, cosH, sinH, i);
   s *= Smax;
-  if (s > 1.0f) /* because of round problems when calculating s and Smax */
-    s = 1.0f;
+  if (s > 1.0) /* because of round problems when calculating s and Smax */
+    s = 1.0;
 
   v = s * cosH;
   u = s * sinH;
 
   /* Inverse of the Parametric equations, using i normalized to 0-1 */
-  *r = i + v/1.5f;
-  *g = i - (v - u*sqrt3)/3.0f;
-  *b = i - (v + u*sqrt3)/3.0f;
+  *r = i + v/1.5;
+  *g = i - (v - u*sqrt3)/3.0;
+  *b = i - (v + u*sqrt3)/3.0;
 
   /* fix round errors */
-  if (*r < 0.0f) *r = 0;
-  if (*g < 0.0f) *g = 0;
-  if (*b < 0.0f) *b = 0;
+  if (*r < 0.0) *r = 0;
+  if (*g < 0.0) *g = 0;
+  if (*b < 0.0) *b = 0;
     
-  if (*r > 1.0f) *r = 1.0f;
-  if (*g > 1.0f) *g = 1.0f;
-  if (*b > 1.0f) *b = 1.0f;
+  if (*r > 1.0) *r = 1.0;
+  if (*g > 1.0) *g = 1.0;
+  if (*b > 1.0) *b = 1.0;
 }
 
 /*********************************************************************************************/
 
-void iupColorRGB2HSI(unsigned char r, unsigned char g, unsigned char b, float *h, float *s, float *i)
+void iupColorRGB2HSI(unsigned char r, unsigned char g, unsigned char b, double *h, double *s, double *i)
 {
-  float fr = iColorReconstruct(r);
-  float fg = iColorReconstruct(g);
-  float fb = iColorReconstruct(b);
+  double fr = iColorReconstruct(r);
+  double fg = iColorReconstruct(g);
+  double fb = iColorReconstruct(b);
   
   iColorRGB2HSI(fr, fg, fb, h, s, i);
 }
 
-void iupColorHSI2RGB(float h, float s, float i, unsigned char *r, unsigned char *g, unsigned char *b)
+void iupColorHSI2RGB(double h, double s, double i, unsigned char *r, unsigned char *g, unsigned char *b)
 {
-  float fr, fg, fb;
+  double fr, fg, fb;
   
   iColorHSI2RGB(h, s, i, &fr, &fg, &fb);
   
@@ -344,11 +344,11 @@ void iupColorHSI2RGB(float h, float s, float i, unsigned char *r, unsigned char 
   *b = iColorQuantize(fb);
 }
 
-int iupStrToHSI(const char *str, float *h, float *s, float *i)
+int iupStrToHSI(const char *str, double *h, double *s, double *i)
 {
-  float fh, fs, fi;
+  double fh, fs, fi;
   if (!str) return 0;
-  if (sscanf(str, "%f %f %f", &fh, &fs, &fi) != 3) return 0;
+  if (sscanf(str, "%lf %lf %lf", &fh, &fs, &fi) != 3) return 0;
   if (fh > 359 || fs > 1 || fi > 1) return 0;
   if (fh < 0 || fs < 0 || fi < 0) return 0;
   *h = fh;
